REFACTOR: PEP8 and SOC
